Add HttpBackend interface
