+++ airframe-http-finagle/src/main/scala/wvlet/airframe/http/finagle/FinagleFilter.scala
@@ -13,46 +13,15 @@
  */
 package wvlet.airframe.http.finagle
 
-import com.twitter.finagle.{Service, SimpleFilter}
 import com.twitter.finagle.http.{Request, Response}
+import com.twitter.finagle.{Service, SimpleFilter}
 import com.twitter.util.Future
-import wvlet.airframe.http.{ControllerProvider, HttpContext, HttpRequestDispatcher, ResponseHandler, Router}
-import wvlet.airframe.http.HttpFilter.HttpFilterFactory
-import wvlet.airframe.http.finagle.FinagleFilter.FinagleFilterFactory
+import wvlet.airframe.http.{ControllerProvider, HttpContext, HttpRequestDispatcher, ResponseHandler}
 
 /**
   * An wrapper of HttpFilter for Finagle backend implementation
   */
-abstract class FinagleFilter extends FinagleFilterFactory.HttpFilterBase
-
-object FinagleFilter {
-
-  object FinagleFilterFactory extends HttpFilterFactory[Request, Response, Future] {
-    override protected def wrapException(e: Throwable): Future[Response] = {
-      Future.exception(e)
-    }
-    override def toFuture[A](a: A): Future[A] = Future.value(a)
-    override def isFutureType(cl: Class[_]): Boolean = {
-      classOf[Future[_]].isAssignableFrom(cl)
-    }
-    override def isRawResponseType(cl: Class[_]): Boolean = {
-      classOf[Response].isAssignableFrom(cl)
-    }
-    override def mapF[A, B](f: Future[A], body: A => B): Future[B] = {
-      f.map(body)
-    }
-    override def newFilter(
-        body: (Request, HttpContext[Request, Response, Future]) => Future[Response]): FinagleFilterFactory.Filter = {
-      new HttpFilterBase {
-        override def apply(request: Request, context: HttpContext[Request, Response, Future]): Future[Response] = {
-          body(request, context)
-        }
-      }
-    }
-  }
-
-  val Identity = FinagleFilterFactory.Identity
-}
+abstract class FinagleFilter extends FinagleBackend.HttpFilterBase
 
 class FinagleRouter(config: FinagleServerConfig,
                     controllerProvider: ControllerProvider,
@@ -60,7 +29,7 @@ class FinagleRouter(config: FinagleServerConfig,
     extends SimpleFilter[Request, Response] {
 
   private val dispatcher =
-    HttpRequestDispatcher.newDispatcher(config.router, controllerProvider, FinagleFilterFactory, responseHandler)
+    HttpRequestDispatcher.newDispatcher(config.router, controllerProvider, FinagleBackend, responseHandler)
 
   override def apply(request: Request, service: Service[Request, Response]): Future[Response] = {
     dispatcher.apply(request, new HttpContext[Request, Response, Future] {
